feat(ui): migrate MenuSection to TypeScript
