+++ packages/ui-components/src/components/MenuSection/index.ts
@@ -3,4 +3,4 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export { MenuSection } from "./MenuSection.component"
+export { MenuSection, type MenuSectionProps } from "./MenuSection.component"
