+++ packages/ui-components/src/index.js
@@ -59,7 +59,7 @@ export { MainContainerInner } from "./components/MainContainerInner/index.js"
 export { MainTabs } from "./components/MainTabs/index.js"
 export { Menu } from "./components/Menu/index.js"
 export { MenuItem } from "./components/MenuItem/index.js"
-export { MenuSection } from "./components/MenuSection/index.js"
+export { MenuSection } from "./components/MenuSection/index.ts"
 export { Message } from "./components/Message/index.ts"
 export { Modal } from "./components/Modal/index.js"
 export { ModalFooter } from "./components/ModalFooter/index.js"
